Change layoutPolicy to deployPolicy for Yum repository closes #380 (#392)
Change layoutPolicy to deployPolicy in the groovy script. Note: on anslible side the var is still named layout_policy as for maven repository (which is the term used in nexus GUI for yum repos anyway)

Working with immutable artifacts is a CICD best practise, therefore the default deploy policy is set to strict. Strict is the Nexus default as well


---------

Co-authored-by: Olivier Clavel <olivier.clavel@gmail.com>